refactor: 컴포넌트 추상화 정리
