+++ src/domain/끄적이는/components/Card/History/style.css.ts
