+++ src/domain/끄적이는/components/Card/components/BookMarkDialog.tsx
@@ -1,0 +1,5 @@
+const BookMarkDialog = () => {
+  return <div></div>;
+};
+
+export default BookMarkDialog;
