+++ src/domain/끄적이는/components/Card/components/SettingDialog.tsx
@@ -1,0 +1,5 @@
+const SettingDialog = () => {
+  return <div></div>;
+};
+
+export default SettingDialog;
