+++ src/domain/끄적이는/components/History/Column/style.css.ts
